fix(plugin-react-router): builtin node server
